Include logging output when running Python code in the docs.
The ipython directive provided by Ipython does not catch the logging
output, which means that the output from Sherpa commands like
sherpa.ui.fit are not included in the documentation. This code
sub-classes the ipython directive so that the sherpa logger writes
to the stream used by the ipython directive.