Improve documentation and build
In Sphinx v1.4.1, re-using the name of a Sphinx extension - as done here
for the ipython directive - causes a warning message to be displayed. It
will be possible to silence this in Sphinx 1.4.2, but for now I have gone
with renaming my overide, to use the name sherpa. At present none of the
examples actually need the functionality here (since it looks like the
logging output is only used by the error routines; if then).

I have looked at using sphinx 1.3 and the ipython directive (rather than
my subclass of it) and do not get sensible behavior with the auto-generated
images, so for now two 'make html' calls are needed.